requirements.txt
flask
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
serpapi

